+++ internal/updater/package_history_other.go
@@ -1,0 +1,7 @@
+//go:build !linux
+
+package updater
+
+func inferPackageHistory(string) (HistoryEntry, bool) {
+	return HistoryEntry{}, false
+}